requirements.txt
customtkinter
Pillow
translitua
pyinstaller